requirements.txt
numpy==1.24.3
pandas==1.5.3
pillow==9.4.0
pytorch==2.0.1
safetensors==0.3.1
torchvision==0.15.2
tqdm==4.65.0
